Chore: 검색 필터링 기능 분리
검색 필터링 기능을 훅으로 분리했습니다.

Related to #1
